[Presentation] Moved the shared image widgets to its own folder
