+++ lib/presentation/artifact/artifact_page.dart
@@ -3,14 +3,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genshindb/application/bloc.dart';
 import 'package:genshindb/generated/l10n.dart';
 import 'package:genshindb/presentation/artifacts/widgets/artifact_stats.dart';
-import 'package:genshindb/presentation/shared/circle_character.dart';
-import 'package:genshindb/presentation/shared/circle_monster.dart';
 import 'package:genshindb/presentation/shared/details/detail_appbar.dart';
 import 'package:genshindb/presentation/shared/details/detail_bottom_portrait_layout.dart';
 import 'package:genshindb/presentation/shared/details/detail_general_card.dart';
 import 'package:genshindb/presentation/shared/details/detail_tab_landscape_layout.dart';
 import 'package:genshindb/presentation/shared/details/detail_top_layout.dart';
 import 'package:genshindb/presentation/shared/extensions/rarity_extensions.dart';
+import 'package:genshindb/presentation/shared/images/circle_character.dart';
+import 'package:genshindb/presentation/shared/images/circle_monster.dart';
 import 'package:genshindb/presentation/shared/item_description_detail.dart';
 import 'package:genshindb/presentation/shared/loading.dart';
 import 'package:genshindb/presentation/shared/scaffold_with_fab.dart';
