+++ lib/presentation/artifacts/widgets/artifact_card.dart
@@ -5,7 +5,7 @@ import 'package:genshindb/domain/models/models.dart';
 import 'package:genshindb/presentation/artifact/artifact_page.dart';
 import 'package:genshindb/presentation/shared/extensions/rarity_extensions.dart';
 import 'package:genshindb/presentation/shared/gradient_card.dart';
-import 'package:genshindb/presentation/shared/rarity.dart';
+import 'package:genshindb/presentation/shared/images/rarity.dart';
 import 'package:genshindb/presentation/shared/styles.dart';
 import 'package:transparent_image/transparent_image.dart';
