+++ lib/presentation/character/widgets/character_detail_talent_ascension_materials_card.dart
@@ -4,9 +4,9 @@ import 'package:genshindb/domain/models/models.dart';
 import 'package:genshindb/generated/l10n.dart';
 import 'package:genshindb/presentation/shared/common_table_cell.dart';
 import 'package:genshindb/presentation/shared/extensions/element_type_extensions.dart';
+import 'package:genshindb/presentation/shared/images/wrapped_ascension_material.dart';
 import 'package:genshindb/presentation/shared/item_description_detail.dart';
 import 'package:genshindb/presentation/shared/styles.dart';
-import 'package:genshindb/presentation/shared/wrapped_ascension_material.dart';
 
 class CharacterDetailTalentAscensionMaterialsCard extends StatelessWidget {
   final ElementType elementType;
@@ -17,14 +17,14 @@ class CharacterDetailTalentAscensionMaterialsCard extends StatelessWidget {
     Key? key,
     required this.elementType,
     required this.talentAscensionMaterials,
-  })   : multiTalentAscensionMaterials = const [],
+  })  : multiTalentAscensionMaterials = const [],
         super(key: key);
 
   const CharacterDetailTalentAscensionMaterialsCard.withMultiTalents({
     Key? key,
     required this.elementType,
     required this.multiTalentAscensionMaterials,
-  })   : talentAscensionMaterials = const [],
+  })  : talentAscensionMaterials = const [],
         super(key: key);
 
   @override
