+++ lib/presentation/characters/widgets/character_card.dart
@@ -5,11 +5,11 @@ import 'package:genshindb/domain/enums/enums.dart';
 import 'package:genshindb/domain/models/models.dart';
 import 'package:genshindb/generated/l10n.dart';
 import 'package:genshindb/presentation/character/character_page.dart';
-import 'package:genshindb/presentation/shared/comingsoon_new_avatar.dart';
-import 'package:genshindb/presentation/shared/element_image.dart';
 import 'package:genshindb/presentation/shared/extensions/element_type_extensions.dart';
 import 'package:genshindb/presentation/shared/extensions/i18n_extensions.dart';
-import 'package:genshindb/presentation/shared/rarity.dart';
+import 'package:genshindb/presentation/shared/images/comingsoon_new_avatar.dart';
+import 'package:genshindb/presentation/shared/images/element_image.dart';
+import 'package:genshindb/presentation/shared/images/rarity.dart';
 import 'package:genshindb/presentation/shared/styles.dart';
 import 'package:genshindb/presentation/shared/utils/toast_utils.dart';
 import 'package:transparent_image/transparent_image.dart';
