+++ lib/presentation/home/widgets/weapon_card_ascension_material.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:genshindb/presentation/home/widgets/ascension_material_item_card.dart';
-import 'package:genshindb/presentation/shared/circle_weapon.dart';
+import 'package:genshindb/presentation/shared/images/circle_weapon.dart';
 
 class WeaponCardAscensionMaterial extends StatelessWidget {
   final String name;
