+++ lib/presentation/shared/images/rarity.dart
