+++ lib/presentation/tierlist/widgets/tierlist_fab.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:genshindb/application/bloc.dart';
-import 'package:genshindb/presentation/shared/circle_character.dart';
+import 'package:genshindb/presentation/shared/images/circle_character.dart';
 import 'package:genshindb/presentation/shared/styles.dart';
 
 class TierListFab extends StatelessWidget {
