+++ lib/presentation/weapon/widgets/weapon_detail_ascension_materials_card.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:genshindb/domain/models/models.dart';
 import 'package:genshindb/generated/l10n.dart';
+import 'package:genshindb/presentation/shared/images/wrapped_ascension_material.dart';
 import 'package:genshindb/presentation/shared/item_description_detail.dart';
 import 'package:genshindb/presentation/shared/styles.dart';
 import 'package:genshindb/presentation/shared/utils/size_utils.dart';
-import 'package:genshindb/presentation/shared/wrapped_ascension_material.dart';
 
 class WeaponDetailAscensionMaterialsCard extends StatelessWidget {
   final Color rarityColor;
