+++ lib/presentation/weapon/widgets/weapon_detail_bottom.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:genshindb/domain/enums/enums.dart';
 import 'package:genshindb/domain/models/models.dart';
 import 'package:genshindb/generated/l10n.dart';
-import 'package:genshindb/presentation/shared/circle_character.dart';
 import 'package:genshindb/presentation/shared/details/detail_bottom_portrait_layout.dart';
 import 'package:genshindb/presentation/shared/details/detail_tab_landscape_layout.dart';
 import 'package:genshindb/presentation/shared/extensions/rarity_extensions.dart';
+import 'package:genshindb/presentation/shared/images/circle_character.dart';
 import 'package:genshindb/presentation/shared/item_description_detail.dart';
 import 'package:genshindb/presentation/shared/styles.dart';
 import 'package:genshindb/presentation/shared/utils/size_utils.dart';
