+++ lib/presentation/weapons/widgets/weapon_card.dart
@@ -6,12 +6,12 @@ import 'package:genshindb/application/bloc.dart';
 import 'package:genshindb/domain/enums/enums.dart';
 import 'package:genshindb/domain/models/models.dart';
 import 'package:genshindb/generated/l10n.dart';
-import 'package:genshindb/presentation/shared/comingsoon_new_avatar.dart';
 import 'package:genshindb/presentation/shared/extensions/i18n_extensions.dart';
 import 'package:genshindb/presentation/shared/extensions/rarity_extensions.dart';
 import 'package:genshindb/presentation/shared/gradient_card.dart';
+import 'package:genshindb/presentation/shared/images/comingsoon_new_avatar.dart';
+import 'package:genshindb/presentation/shared/images/rarity.dart';
 import 'package:genshindb/presentation/shared/loading.dart';
-import 'package:genshindb/presentation/shared/rarity.dart';
 import 'package:genshindb/presentation/shared/styles.dart';
 import 'package:genshindb/presentation/weapon/weapon_page.dart';
 import 'package:transparent_image/transparent_image.dart';
